Added deleting photo feature

+++ app/src/main/java/com/njscky/mapcollect/business/photo/DisplayPhotoActivity.java
@@ -2,15 +2,11 @@ package com.njscky.mapcollect.business.photo;
 
 import android.app.Activity;
 import android.content.Intent;
+import android.net.Uri;
 import android.os.Bundle;
 import android.util.Log;
 import android.view.Menu;
 import android.view.MenuItem;
-
-import androidx.annotation.Nullable;
-import androidx.appcompat.app.AppCompatActivity;
-import androidx.appcompat.widget.Toolbar;
-import androidx.viewpager.widget.ViewPager;
 
 import com.google.android.material.tabs.TabLayout;
 import com.njscky.mapcollect.R;
@@ -21,8 +17,12 @@ import com.njscky.mapcollect.util.AppExecutors;
 
 import java.io.File;
 import java.util.ArrayList;
-import java.util.List;
 
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.AlertDialog;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
+import androidx.viewpager.widget.ViewPager;
 import butterknife.BindView;
 import butterknife.ButterKnife;
 
@@ -39,23 +39,40 @@ public class DisplayPhotoActivity extends AppCompatActivity {
     @BindView(R.id.tabLayout)
     TabLayout tabLayout;
     PhotoPagerAdapter adapter;
-    private List<PhotoJCJ> photos;
+    private ArrayList<PhotoJCJ> photos;
     private int current;
     private PhotoJCJDao photoJCJDao;
 
     public static void start(Activity activity, ArrayList<PhotoJCJ> photos, int current) {
         Intent intent = new Intent(activity, DisplayPhotoActivity.class);
-
         intent.putParcelableArrayListExtra("photos", photos);
         intent.putExtra("current", current);
-
         activity.startActivity(intent);
+    }
+
+    public static void startForResult(Activity activity, ArrayList<PhotoJCJ> photos, int current, int requestCode) {
+        Intent intent = new Intent(activity, DisplayPhotoActivity.class);
+        intent.putParcelableArrayListExtra("photos", photos);
+        intent.putExtra("current", current);
+        activity.startActivityForResult(intent, requestCode);
     }
 
     @Override
     public boolean onCreateOptionsMenu(Menu menu) {
-//        getMenuInflater().inflate(R.menu.menu_display_photo, menu);
+        getMenuInflater().inflate(R.menu.menu_display_photo, menu);
         return super.onCreateOptionsMenu(menu);
+    }
+
+    @Override
+    public boolean onPrepareOptionsMenu(Menu menu) {
+        for (int i = 0; i < menu.size(); i++) {
+            MenuItem menuItem = menu.getItem(i);
+            if (menuItem.getItemId() == R.id.action_delete) {
+                menuItem.setEnabled(photos != null && !photos.isEmpty());
+            }
+        }
+        return super.onPrepareOptionsMenu(menu);
+
     }
 
     @Override
@@ -65,18 +82,29 @@ public class DisplayPhotoActivity extends AppCompatActivity {
         extractExtras(getIntent());
         ButterKnife.bind(this);
         setSupportActionBar(toolbar);
-        toolbar.setOnMenuItemClickListener(new Toolbar.OnMenuItemClickListener() {
-            @Override
-            public boolean onMenuItemClick(MenuItem item) {
-                switch (item.getItemId()) {
-                    case R.id.action_delete:
-                        deletePhoto();
-                        break;
-                }
-                return false;
-            }
+        toolbar.setNavigationOnClickListener((view) -> {
+            onBackPressed();
         });
+        toolbar.setOnMenuItemClickListener(item -> {
+            switch (item.getItemId()) {
+                case R.id.action_delete:
+                    if (photos != null && !photos.isEmpty()) {
+                        new AlertDialog.Builder(DisplayPhotoActivity.this)
+                                .setTitle(R.string.dialog_title)
+                                .setMessage(R.string.dialog_delete_photo)
+                                .setNegativeButton(R.string.cancel, (dialog, which) -> {
 
+                                })
+                                .setPositiveButton(R.string.confirm, (dialog, which) -> {
+                                    deletePhoto();
+
+                                })
+                                .show();
+                    }
+                    break;
+            }
+            return false;
+        });
         photoJCJDao = DbManager.getInstance(this).getDaoSession().getPhotoJCJDao();
         adapter = new PhotoPagerAdapter(photos);
         vpPhotos.setAdapter(adapter);
@@ -98,6 +126,7 @@ public class DisplayPhotoActivity extends AppCompatActivity {
             public void onPageScrollStateChanged(int state) {
 
             }
+
         });
         tabLayout.setupWithViewPager(vpPhotos, true);
 
@@ -113,12 +142,14 @@ public class DisplayPhotoActivity extends AppCompatActivity {
                 File targetFile = new File(target.ZPLJ);
                 if (targetFile.exists() && targetFile.canWrite()) {
                     if (targetFile.delete()) {
+                        sendBroadcast(new Intent(Intent.ACTION_MEDIA_SCANNER_SCAN_FILE, Uri.fromFile(targetFile)));
                         if (target.ID != null) {
                             photoJCJDao.delete(target);
                         }
                         AppExecutors.MAIN.execute(() -> {
                             photos.remove(current);
                             adapter.notifyDataSetChanged();
+                            supportInvalidateOptionsMenu();
                         });
                         Log.i(TAG, "delete successful: " + target.ZPLJ);
                     }
@@ -127,6 +158,14 @@ public class DisplayPhotoActivity extends AppCompatActivity {
         });
 
 
+    }
+
+    @Override
+    public void onBackPressed() {
+        Intent data = new Intent();
+        data.putParcelableArrayListExtra("photos", photos);
+        setResult(RESULT_OK, data);
+        super.onBackPressed();
     }
 
     private void extractExtras(Intent intent) {
